Added FridayP3.py and changed the previous file name to FridayP2.py to match python versions
